refactor(http-benchmarks): refactor currentInversifyExpress to use K6
